ASoC: msm: qdsp6v2: Fix impossible check
We assign the long type here to bool, so it can only be true or
false, not negative. Replace the impossible check with the
correct one, which makes sure that only 0 or a positive integer
is used as an argument here. This also removes a warning on newer
compilers.

We already had a fix in place for that, but this solution is a
bit cleaner.

Change-Id: If31c160a241b7389038288bfee3986a060fcc959
CRs-Fixed: 1099676
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Mohammad Hasan Keramat J <ikeramat@protonmail.com>