Restructure the battery detection code, and other improvements.
Instead of copying a lot of ifdefs everywhere we're going to
pass battery data, it makes more sense to isolate those to a
special file that handles those things and can be expaned a lot
more in the future. Not only that, but the battery profile for
the G5 was also updated and should be working now.