Small improvement to the debug layout.
As @stendro pointed out, some -DDEBUG flags need to be applied
to all subdirs where the Makefile is located, that's the case for
usb for example. Only the important ones were turned into subdir
flags while others were in-depth enough to not need them, as well
as the subdirs that Lg's trio didn't even use. As for the file
omissions to reduce log spam, i prefer to let them in seeing as
power_supply debug also tendds to spam a lot of _changed() calls,
and we also won't be as impacted by them unless we enable all at
the same time so we can keep the amount of changes to a minimum.