dts: lge: Updating the fuel gauge and smbcharger props.
About time we started looking at those props again, with updates
to temps, fg thresholds and somc current tables.

Besides those adjustments, we're now going to use the added somc
props to better handle temp mitigation on the phones, which is
confirmed to massively improve the charging speed on the V20.