msm: camera: Adjust the LGE ifdefs in msmb_camera.h
In 4.4, qcom added a few more group ids for camera subdevices,
which may conflict with LGE's 3.18 group ids. So the MACH_LGE
ifdef is now adjusted to remove them in case one of those
devices is the build target. The assumption is that if we didn't
have those new groups in 3.18, our camera doesn't need or use them